requirements.txt
discord.py
flask
aiohttp
supabase
python-dotenv
